requirements.txt
# OpenAI Whisper for GPU
faster-whisper
torch
torchaudio
numpy
pynput
sounddevice
pyperclip
pyyaml
scipy

# Optional but recommended
ffmpeg-python

# GUI & Packaging
PyQt6
pyinstaller
